requirements.txt
# =============================================================================
# Marketing Mix Modelling — Python Requirements
# Python 3.10+
# =============================================================================

# --- Core data manipulation ---
pandas>=2.0.0
numpy>=1.26.0

# --- Statistical modelling ---
statsmodels>=0.14.0
scikit-learn>=1.4.0
scipy>=1.12.0

# --- Bayesian MMM ---
pymc>=5.10.0
pymc-marketing>=0.5.0

# --- Google Meridian ---
# meridian  # Install separately: pip install google-meridian
# jax[cpu]>=0.4.20
# tensorflow-probability>=0.23.0

# --- Bayesian diagnostics ---
arviz>=0.18.0

# --- Accelerated numerics (Meridian backend) ---
# jax>=0.4.20
# jaxlib>=0.4.20

# --- Visualisation ---
plotly>=5.20.0
matplotlib>=3.8.0
seaborn>=0.13.0
kaleido>=0.2.1          # Static Plotly image export

# --- Dashboard ---
streamlit>=1.32.0

# --- Reporting ---
python-pptx>=0.6.23

# --- Notebooks ---
jupyter>=1.0.0
jupyterlab>=4.1.0
ipykernel>=6.29.0
nbformat>=5.9.0

# --- Data download ---
kaggle>=1.6.0

# --- Config ---
pyyaml>=6.0.1
python-dotenv>=1.0.0

# --- Utilities ---
tqdm>=4.66.0
joblib>=1.3.0
